extract cloud service from track service

+++ api/v1/track.py
@@ -2,7 +2,7 @@ from typing import Sequence
 from pathlib import Path
 from fastapi import APIRouter,HTTPException,status,Depends,Query,UploadFile,File,Form
 from schemas import TrackDownloadSchema,TrackSchema,TrackUploadSchema,TrakUpdateSchema,UserSchema
-from services import TrackService,get_track_service,get_current_user
+from services import TrackService,get_track_service,get_current_user,BackBlazeB2Service,get_backblazeb2_service
 from settings import ENVIRONMENT
 
 router = APIRouter(prefix='/tracks',tags=['tracks'])
@@ -17,7 +17,8 @@ async def upload_track(
     author_name:str,
     data:UploadFile = File(...),
     current_user:UserSchema=Depends(get_current_user),
-    track_service:TrackService=Depends(get_track_service)
+    track_service:TrackService=Depends(get_track_service),
+    cloud_service:BackBlazeB2Service=Depends(get_backblazeb2_service)
 ):
     if not data.filename:
         raise HTTPException(
@@ -27,7 +28,9 @@ async def upload_track(
     extension = Path(data.filename).suffix
     track = TrackUploadSchema(name=f'{track_name}{extension}',author_name=author_name)
     track_data = await data.read()
-    return await track_service.create(track,data=track_data)
+    cloud_response = await cloud_service.upload_file(track_data,track.name)
+
+    return await track_service.create(track,id=cloud_response.id,size=cloud_response.size)
 
 @router.get(
     '',
@@ -68,13 +71,20 @@ async def get_track(
 )
 async def get_track_url(
     track_id:str,
-    service:TrackService=Depends(get_track_service)
+    service:TrackService=Depends(get_track_service),
+    cloud_service:BackBlazeB2Service=Depends(get_backblazeb2_service)
 ):
-    db_track = await service.get_track_url(track_id)
+    db_track = await service.get_by_id(track_id)
     if not db_track:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f'No track with id "{track_id}" was found'
+        )
+    db_track = await cloud_service.get_file(db_track)
+    if not db_track:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail='The file was not found in the cloud'
         )
     return db_track
 
@@ -85,8 +95,21 @@ async def get_track_url(
 async def delete(
     track_id:str,
     service:TrackService=Depends(get_track_service),
-    current_user:UserSchema=Depends(get_current_user)
+    current_user:UserSchema=Depends(get_current_user),
+    cloud_service:BackBlazeB2Service=Depends(get_backblazeb2_service)
 ):
+    db_track = await service.get_by_id(track_id)
+    if not db_track:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f'No track with id {track_id} was found'
+        )
+    cloud_deleted = await cloud_service.remove_file(track_id,db_track.name)
+    if not cloud_deleted:
+        raise HTTPException(
+            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
+            detail='The track was not deleted from cloud'
+        )
     deleted = await service.delete(track_id)
     if not deleted:
         raise HTTPException(
